+++ src/test/java/step5/domain/CarTest.java
@@ -1,4 +1,4 @@
-package step4.domain;
+package step5.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
